refactor input_pipeline.py and fix test
