+++ MaxText/input_pipeline/_tfds_data_processing.py
@@ -23,8 +23,6 @@ import ml_collections
 import tensorflow as tf
 import tensorflow_datasets as tfds
 import jax
-import jax.numpy as jnp
-from jax.sharding import PartitionSpec as P
 
 import tokenizer
 import multihost_dataloading
@@ -269,77 +267,3 @@ def preprocess_dataset(config: ml_collections.ConfigDict,
 
   return train_iter, eval_iter, predict_iter, sp_tokenizer
 
-
-def make_c4_train_iterator_and_tokenizer(config, mesh, add_bos, add_eos):
-  """ Make train iterator and tokenizer for C4 dataset"""
-  read_config = tfds.ReadConfig(
-    shuffle_seed = config.data_shuffle_seed,
-  )
-  train_ds, eval_ds = get_datasets(
-    config=config,
-    read_config = read_config,
-  )
-  train_iter, _, _, sp_tokenizer = preprocess_dataset(
-    config,
-    mesh,
-    train_ds, eval_ds,
-    vocab_path=os.path.join(config.assets_path, config.vocab_relative_path),
-    data_shuffle_seed = config.data_shuffle_seed,
-    add_bos = add_bos,
-    add_eos = add_eos
-  )
-  return train_iter, sp_tokenizer
-
-class SyntheticDataIterator():
-  """Creates a synthetic data iterator for performance testing work"""
-  def __init__(self, config, mesh):
-    self.mesh = mesh
-    self.config = config
-    data_pspec = P(*config.data_sharding)
-    data_pspec_shardings = jax.tree_map(
-        lambda p: jax.sharding.NamedSharding(mesh, p), data_pspec)
-    self.data_generator = jax.jit(SyntheticDataIterator.raw_generate_synthetic_data,
-        out_shardings=data_pspec_shardings,
-        static_argnums=0)
-  def __call__(self):
-    with self.mesh:
-      return self.data_generator(self.config)
-
-  @staticmethod
-  def raw_generate_synthetic_data(config):
-    """Generates a single batch of syntehtic data"""
-    output = {}
-    output['inputs'] = jax.numpy.zeros( (config.global_batch_size_to_load, config.max_target_length),
-                                       dtype=jax.numpy.int32)
-    output['inputs_position'] = jax.numpy.zeros( (config.global_batch_size_to_load, config.max_target_length),
-                                                dtype=jax.numpy.int32)
-    output['inputs_segmentation'] = jax.numpy.ones( (config.global_batch_size_to_load, config.max_target_length),
-                                                   dtype=jax.numpy.int32)
-    output['targets'] = jax.numpy.zeros( (config.global_batch_size_to_load, config.max_target_length),
-                                        dtype=jax.numpy.int32)
-    output['targets_position'] = jax.numpy.zeros( (config.global_batch_size_to_load, config.max_target_length),
-                                                 dtype=jax.numpy.int32)
-    output['targets_segmentation'] = jax.numpy.ones( (config.global_batch_size_to_load, config.max_target_length),
-                                                    dtype=jax.numpy.int32)
-    return output
-
-def create_data_iterator_with_tokenizer(config, mesh, add_bos = True, add_eos = True):
-  if config.dataset_type == "synthetic":
-    return SyntheticDataIterator(config, mesh), None
-  elif config.dataset_type == "c4":
-    return make_c4_train_iterator_and_tokenizer(config, mesh, add_bos, add_eos)
-  else:
-    assert False, "dataset type not implemented"
-
-def get_shaped_batch(config):
-  """ Return the shape of the batch - this is what eval_shape would return for the
-  output of create_data_iterator_with_tokenizer, but eval_shape doesn't work, see b/306901078."""
-  batch_shape = (config.global_batch_size_to_load, config.max_target_length)
-  shaped_batch = {}
-  shaped_batch['inputs'] = jax.ShapeDtypeStruct(batch_shape, jnp.int32)
-  shaped_batch['inputs_position'] = jax.ShapeDtypeStruct(batch_shape, jnp.int32)
-  shaped_batch['inputs_segmentation'] = jax.ShapeDtypeStruct(batch_shape, jnp.int32)
-  shaped_batch['targets'] = jax.ShapeDtypeStruct(batch_shape, jnp.int32)
-  shaped_batch['targets_position'] = jax.ShapeDtypeStruct(batch_shape, jnp.int32)
-  shaped_batch['targets_segmentation'] = jax.ShapeDtypeStruct(batch_shape, jnp.int32)
-  return shaped_batch
