+++ MaxText/pytest.ini
@@ -3,6 +3,6 @@
 testpaths =
     tests
 python_files = *_test.py
-addopts = -rf --import-mode=importlib --ignore=tests/input_pipeline_test.py --ignore=tests/profiler_test.py --ignore=tests/train_smoke_test.py
+addopts = -rf --import-mode=importlib --ignore=tests/profiler_test.py --ignore=tests/train_smoke_test.py
 markers = 
     tpu: marks tests to be run on TPU
